Remove support for personal sign
